# apps/web/openapi.json
{
  "components": {
    "schemas": {
      "ApplyRequest": {
        "properties": {
          "confirm": {
            "default": false,
            "description": "must be true: a write is never the default outcome of a request",
            "title": "Confirm",
            "type": "boolean"
          },
          "entity_id": {
            "description": "id of the archived entity to map from",
            "title": "Entity Id",
            "type": "integer"
          },
          "entity_kind": {
            "anyOf": [
              {
                "enum": [
                  "artist",
                  "album",
                  "track"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "defaults to the kind implied by the Jellyfin item",
            "title": "Entity Kind"
          },
          "expected_etag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "the etag the diff was prepared against; a mismatch is refused",
            "title": "Expected Etag"
          },
          "fields": {
            "anyOf": [
              {
                "items": {
                  "type": "string"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "description": "fields to write; omit to use the plan's default selection",
            "title": "Fields"
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/FieldPolicyOverride"
            },
            "title": "Overrides",
            "type": "array"
          },
          "tag_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TagPolicyRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "entity_id"
        ],
        "title": "ApplyRequest",
        "type": "object"
      },
      "ApplyResponse": {
        "properties": {
          "applied": {
            "items": {
              "type": "string"
            },
            "title": "Applied",
            "type": "array"
          },
          "duration_ms": {
            "title": "Duration Ms",
            "type": "integer"
          },
          "etag_after": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etag After"
          },
          "etag_before": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etag Before"
          },
          "idempotency_key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Idempotency Key"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "snapshot_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Snapshot Id"
          },
          "state": {
            "anyOf": [
              {
                "additionalProperties": true,
                "type": "object"
              },
              {
                "type": "null"
              }
            ],
            "title": "State"
          },
          "unchanged": {
            "items": {
              "type": "string"
            },
            "title": "Unchanged",
            "type": "array"
          }
        },
        "required": [
          "item_id",
          "kind",
          "applied",
          "unchanged",
          "duration_ms"
        ],
        "title": "ApplyResponse",
        "type": "object"
      },
...
      "CandidateSummary": {
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "mbid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "response_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Response Id"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "name"
        ],
        "title": "CandidateSummary",
        "type": "object"
      },
      "CandidateTagView": {
        "properties": {
          "count": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Count"
          },
          "name": {
            "title": "Name",
            "type": "string"
          }
        },
        "required": [
          "name"
        ],
        "title": "CandidateTagView",
        "type": "object"
      },
      "CandidateView": {
        "properties": {
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceView"
          },
          "duration_ms": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Ms"
          },
          "entity_id": {
            "title": "Entity Id",
            "type": "integer"
          },
          "entity_kind": {
            "enum": [
              "artist",
              "album",
              "track"
            ],
            "title": "Entity Kind",
            "type": "string"
          },
          "has_overview": {
            "default": false,
            "title": "Has Overview",
            "type": "boolean"
          },
          "identity": {
            "title": "Identity",
            "type": "string"
          },
          "last_seen_at": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Last Seen At"
          },
          "latest_response_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Latest Response Id"
          },
          "listeners": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Listeners"
          },
          "matched_on": {
            "enum": [
              "mbid",
              "name"
            ],
            "title": "Matched On",
            "type": "string"
          },
          "mbid": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Mbid"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "playcount": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Playcount"
          },
          "tags": {
            "items": {
              "$ref": "#/components/schemas/CandidateTagView"
            },
            "title": "Tags",
            "type": "array"
          },
          "url": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Url"
          },
          "year": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Year"
          }
        },
        "required": [
          "entity_kind",
          "entity_id",
          "identity",
          "name",
          "matched_on",
          "confidence"
        ],
        "title": "CandidateView",
        "type": "object"
      },
      "CandidatesRequest": {
        "properties": {
          "limit": {
            "default": 5,
            "maximum": 25.0,
            "minimum": 1.0,
            "title": "Limit",
            "type": "integer"
          },
          "refresh": {
            "default": false,
            "description": "Reserved: fetching from Last.fm is not part of the read-only diff path.",
            "title": "Refresh",
            "type": "boolean"
          }
        },
        "title": "CandidatesRequest",
        "type": "object"
      },
      "CandidatesResponse": {
        "properties": {
          "candidates": {
            "items": {
              "$ref": "#/components/schemas/CandidateView"
            },
            "title": "Candidates",
            "type": "array"
          },
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "etag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etag"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "note": {
            "title": "Note",
            "type": "string"
          }
        },
        "required": [
          "item_id",
          "kind",
          "name",
          "candidates",
          "count",
          "note"
        ],
        "title": "CandidatesResponse",
        "type": "object"
      },
...
      "DiffRequest": {
        "properties": {
          "entity_id": {
            "description": "id of the archived entity to map from",
            "title": "Entity Id",
            "type": "integer"
          },
          "entity_kind": {
            "anyOf": [
              {
                "enum": [
                  "artist",
                  "album",
                  "track"
                ],
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "description": "defaults to the kind implied by the Jellyfin item",
            "title": "Entity Kind"
          },
          "overrides": {
            "items": {
              "$ref": "#/components/schemas/FieldPolicyOverride"
            },
            "title": "Overrides",
            "type": "array"
          },
          "tag_policy": {
            "anyOf": [
              {
                "$ref": "#/components/schemas/TagPolicyRequest"
              },
              {
                "type": "null"
              }
            ]
          }
        },
        "required": [
          "entity_id"
        ],
        "title": "DiffRequest",
        "type": "object"
      },
      "DiffResponse": {
        "properties": {
          "candidate": {
            "$ref": "#/components/schemas/CandidateSummary"
          },
          "changes": {
            "items": {
              "$ref": "#/components/schemas/FieldChangeView"
            },
            "title": "Changes",
            "type": "array"
          },
          "confidence": {
            "$ref": "#/components/schemas/ConfidenceView"
          },
          "default_selection": {
            "items": {
              "type": "string"
            },
            "title": "Default Selection",
            "type": "array"
          },
          "etag": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Etag"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "kind": {
            "title": "Kind",
            "type": "string"
          },
          "locked_fields": {
            "items": {
              "type": "string"
            },
            "title": "Locked Fields",
            "type": "array"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "withheld": {
            "items": {
              "$ref": "#/components/schemas/FieldChangeView"
            },
            "title": "Withheld",
            "type": "array"
          }
        },
        "required": [
          "item_id",
          "kind",
          "name",
          "confidence",
          "candidate",
          "changes",
          "withheld",
          "default_selection"
        ],
        "title": "DiffResponse",
        "type": "object"
      },
...
      "FieldPolicyOverride": {
        "description": "Per-field mode for one request, so a single edit need not change configuration.",
        "properties": {
          "field": {
            "title": "Field",
            "type": "string"
          },
          "mode": {
            "enum": [
              "keep_existing",
              "fill_if_empty",
              "replace",
              "merge"
            ],
            "title": "Mode",
            "type": "string"
          }
        },
        "required": [
          "field",
          "mode"
        ],
        "title": "FieldPolicyOverride",
        "type": "object"
      },
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
...
      "SnapshotListResponse": {
        "properties": {
          "count": {
            "title": "Count",
            "type": "integer"
          },
          "item_id": {
            "title": "Item Id",
            "type": "string"
          },
          "snapshots": {
            "items": {
              "$ref": "#/components/schemas/SnapshotSummary"
            },
            "title": "Snapshots",
            "type": "array"
          }
        },
        "required": [
          "item_id",
          "snapshots",
          "count"
        ],
        "title": "SnapshotListResponse",
        "type": "object"
      },
...
      "TagPolicyRequest": {
        "properties": {
          "extra_blacklist": {
            "items": {
              "type": "string"
            },
            "title": "Extra Blacklist",
            "type": "array"
          },
          "genre_limit": {
            "default": 5,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Genre Limit",
            "type": "integer"
          },
          "min_count": {
            "default": 0,
            "minimum": 0.0,
            "title": "Min Count",
            "type": "integer"
          },
          "style_limit": {
            "default": 10,
            "maximum": 50.0,
            "minimum": 0.0,
            "title": "Style Limit",
            "type": "integer"
          }
        },
        "title": "TagPolicyRequest",
        "type": "object"
      },
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    }
  },
  "info": {
    "description": "Browse a Jellyfin music library, read artist/album/track metadata from the\nLast.fm API, keep a persistent incremental local archive of every Last.fm\nresponse, review a field-level diff, and apply the accepted changes back to\nJellyfin.\n",
    "title": "metaedit",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/items/{item_id}/apply": {
      "post": {
        "description": "Apply a reviewed change set.\n\n``confirm`` must be true. A write is never the default outcome of a request, and\nan omitted selection means the plan's default -- which is empty for anything that\nneeds review, so an unreviewed match writes nothing.",
        "operationId": "item_apply_api_items__item_id__apply_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "header",
            "name": "Idempotency-Key",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Idempotency-Key"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApplyRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Item Apply",
        "tags": [
          "editing"
        ]
      }
    },
    "/api/items/{item_id}/candidates": {
      "post": {
        "description": "Archived Last.fm entities that could be this item, best match first.\n\nReads only; nothing is fetched and nothing is written.",
        "operationId": "item_candidates_api_items__item_id__candidates_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CandidatesRequest",
                "default": {
                  "limit": 5,
                  "refresh": false
                }
              }
            }
          }
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/CandidatesResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Item Candidates",
        "tags": [
          "editing"
        ]
      }
    },
    "/api/items/{item_id}/diff": {
      "post": {
        "description": "Preview the exact changes applying this candidate would make.\n\nEverything is computed and returned; nothing is written. The response includes\nthe withheld fields and the reasons, so a policy can be debugged rather than\nguessed at.",
        "operationId": "item_diff_api_items__item_id__diff_post",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/DiffRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DiffResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Item Diff",
        "tags": [
          "editing"
        ]
      }
    },
...
    "/api/items/{item_id}/snapshots": {
      "get": {
        "description": "Snapshot history, newest first, so any applied change can be undone.",
        "operationId": "item_snapshots_api_items__item_id__snapshots_get",
        "parameters": [
          {
            "in": "path",
            "name": "item_id",
            "required": true,
            "schema": {
              "title": "Item Id",
              "type": "string"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SnapshotListResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Item Snapshots",
        "tags": [
          "editing"
        ]
      }
    },
...
    "/api/snapshots/{snapshot_id}/revert": {
      "post": {
        "description": "Restore the values captured by a snapshot.\n\nA revert is an ordinary write built from the snapshot's fields, and it is itself\nsnapshotted -- so undoing an undo works and history is never mutated.",
        "operationId": "revert_api_snapshots__snapshot_id__revert_post",
        "parameters": [
          {
            "in": "path",
            "name": "snapshot_id",
            "required": true,
            "schema": {
              "title": "Snapshot Id",
              "type": "integer"
            }
          },
          {
            "description": "must be true",
            "in": "query",
            "name": "confirm",
            "required": false,
            "schema": {
              "default": false,
              "description": "must be true",
              "title": "Confirm",
              "type": "boolean"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApplyResponse"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "summary": "Revert",
        "tags": [
          "editing"
        ]
      }
    }
  }
}